+++ frontend/src/pages/ReviewWritingPage/layout/index.tsx
@@ -1,0 +1,1 @@
+export { default as QnABoxLayout } from './QnABoxLayout';
